Errors handler architectur
Implemented a custom errors handler for the front to display errors from
the API.
A dispatcher is call first to return the abstract `ApiErrorHander`
object, that have many children that can inherit from it.
For example, when patching a review , a list of `Api400Handler` objects are
return if a badrequest error occured.

Implemented a store dispatcher for the errors also, that will help
propagate the error into sub component. An implementation have been make
when adding a review. The component `AddAReview` listen to the store to
display a bunch of erros related to the "grade" field, and the erros are
propagated from the `MovieView` view by passing to the store.

A documentation will be write soon to explain how to use this new
architecture and help you control your error messages.